server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Wallup - AI wall color changer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.132.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "huggingface-hub>=1.4.1",
]
